requirements.txt
asgiref==3.3.4
Django==3.0.5
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
mysqlclient==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
